Update with Collision Handler
The Collision handler script has been updated, primarily, to allow the rocket to be destroyed when the player lands somewhere that isn't the finish line and to load the next level if they successfully land on the finish line.

One new scene has been created, some game objects have been prefabbed to make this easier for new levels in the future.